+++ src/Indexing/DirectoryIndexRequestDemuxer.php
@@ -82,7 +82,7 @@ class DirectoryIndexRequestDemuxer
         JsonRpcResponseSenderInterface $jsonRpcResponseSender
     ): void {
         $request = new JsonRpcRequest(null, 'reindex', [
-            'source'    => [$fileInfo->getPathname()],
+            'uris'      => [$fileInfo->getPathname()],
             'exclude'   => $globsToExclude,
             'extension' => $extensionsToIndex,
         ]);