+++ tests/Integration/UserInterface/Command/LocalizeTypeCommandTest.php
@@ -4,6 +4,8 @@ namespace Serenata\Tests\Integration\UserInterface\Command;
 
 use Serenata\Analysis\Visiting\UseStatementKind;
 
+use Serenata\Common\Position;
+
 use Serenata\Indexing\FileNotFoundStorageException;
 use Serenata\Tests\Integration\AbstractIntegrationTest;
 
@@ -20,19 +22,19 @@ class LocalizeTypeCommandTest extends AbstractIntegrationTest
 
         $command = $this->container->get('localizeTypeCommand');
 
-        static::assertSame('C', $command->localizeType('C', $path, 1, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('\C', $command->localizeType('\C', $path, 5, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('C', $command->localizeType('\A\C', $path, 5, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('C', $command->localizeType('\B\C', $path, 10, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('DateTime', $command->localizeType('\B\DateTime', $path, 10, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('DateTime', $command->localizeType('\DateTime', $path, 11, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('DateTime', $command->localizeType('DateTime', $path, 12, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('DateTime', $command->localizeType('\DateTime', $path, 12, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('D\Test', $command->localizeType('\C\D\Test', $path, 13, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('E', $command->localizeType('\C\D\E', $path, 14, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('H', $command->localizeType('\F\G\H', $path, 16, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('SOME_CONSTANT', $command->localizeType('\A\SOME_CONSTANT', $path, 18, UseStatementKind::TYPE_CONSTANT));
-        static::assertSame('some_function', $command->localizeType('\A\some_function', $path, 18, UseStatementKind::TYPE_FUNCTION));
+        static::assertSame('C', $command->localizeType('C', $path, new Position(0, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('\C', $command->localizeType('\C', $path, new Position(4, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('C', $command->localizeType('\A\C', $path, new Position(4, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('C', $command->localizeType('\B\C', $path, new Position(9, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('DateTime', $command->localizeType('\B\DateTime', $path, new Position(9, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('DateTime', $command->localizeType('\DateTime', $path, new Position(10, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('DateTime', $command->localizeType('DateTime', $path, new Position(11, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('DateTime', $command->localizeType('\DateTime', $path, new Position(11, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('D\Test', $command->localizeType('\C\D\Test', $path, new Position(12, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('E', $command->localizeType('\C\D\E', $path, new Position(13, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('H', $command->localizeType('\F\G\H', $path, new Position(15, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('SOME_CONSTANT', $command->localizeType('\A\SOME_CONSTANT', $path, new Position(17, 0), UseStatementKind::TYPE_CONSTANT));
+        static::assertSame('some_function', $command->localizeType('\A\some_function', $path, new Position(17, 0), UseStatementKind::TYPE_FUNCTION));
     }
 
     /**
@@ -46,9 +48,9 @@ class LocalizeTypeCommandTest extends AbstractIntegrationTest
 
         $command = $this->container->get('localizeTypeCommand');
 
-        static::assertSame('\C\D\Test', $command->localizeType('\C\D\Test', $path, 13, UseStatementKind::TYPE_CONSTANT));
-        static::assertSame('\SOME_CONSTANT', $command->localizeType('\SOME_CONSTANT', $path, 18, UseStatementKind::TYPE_CLASSLIKE));
-        static::assertSame('\some_function', $command->localizeType('\some_function', $path, 18, UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('\C\D\Test', $command->localizeType('\C\D\Test', $path, new Position(12, 0), UseStatementKind::TYPE_CONSTANT));
+        static::assertSame('\SOME_CONSTANT', $command->localizeType('\SOME_CONSTANT', $path, new Position(17, 0), UseStatementKind::TYPE_CLASSLIKE));
+        static::assertSame('\some_function', $command->localizeType('\some_function', $path, new Position(17, 0), UseStatementKind::TYPE_CLASSLIKE));
     }
 
     /**
@@ -60,6 +62,6 @@ class LocalizeTypeCommandTest extends AbstractIntegrationTest
 
         $this->expectException(FileNotFoundStorageException::class);
 
-        $command->localizeType('A', 'DoesNotExist.phpt', 1, UseStatementKind::TYPE_CLASSLIKE);
+        $command->localizeType('A', 'DoesNotExist.phpt', new Position(0, 1), UseStatementKind::TYPE_CLASSLIKE);
     }
 }